--- framework/Models/PostProcessors/validationAlgorithms/__init__.py
@@ -1,24 +1,0 @@
-# Copyright 2017 Battelle Energy Alliance, LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-  The Validation module includes algorithms for deploying validation method in RAVEN.
-
-  Created on April, 2021
-  @author: alfoa
-"""
-
-# These lines ensure that we do not have to do something like:
-#from .ValidationBase import ValidationBase
-
-from .Factory import factory
